split the mangaer.go file to multiple smaller files
